Update item_entry.cs, Program.cs, and 88 more files...

+++ POSPRA.Application/Services/FileRecordService/IFileRecordService.cs
@@ -6,7 +6,7 @@ namespace POSPRA.Application.Services.FileRecordService
     public interface IFileRecordService
     {
         Task<int> CreateAsync(long posId, string encryptedData, string invoiceNumber);
-        Task<ApiResponse<FileRecordDto>> GetByInvoiceIdAsync(long invoiceId);
+        Task<ApiResponse<FileRecordDto>> GetByInvoiceIdAsync(int invoiceId);
         Task<ApiResponse<List<FileRecordDto>>> GetAllAsync();
         Task<ApiResponse<List<FileRecordDto>>> GetAllUnsyncedAsync();
         Task<ApiResponse<FileRecordDto>> UpdateFileRecordAsync(FileRecordDto fileRecordDto);
